+++ ginger_libmt94x/tests/test_transfer_failed_codes.py
@@ -1,9 +1,9 @@
 from unittest import TestCase
 
-from ginger.libmt94x.transfer_failed_codes import InvalidTransferFailedCodeError
-from ginger.libmt94x.transfer_failed_codes import TransferFailed
-from ginger.libmt94x.transfer_failed_codes import TransferFailedMisc
-from ginger.libmt94x.transfer_failed_codes import TransferFailedSEPA
+from ginger_libmt94x.transfer_failed_codes import InvalidTransferFailedCodeError
+from ginger_libmt94x.transfer_failed_codes import TransferFailed
+from ginger_libmt94x.transfer_failed_codes import TransferFailedMisc
+from ginger_libmt94x.transfer_failed_codes import TransferFailedSEPA
 
 
 class TransferFailedTests(TestCase):