+++ ginger_libmt94x/tests/test_writer_doc_ibp_unstructured.py
@@ -4,21 +4,21 @@ from decimal import Decimal
 from unittest import TestCase
 import os
 
-from ginger.libmt94x.document import Mt940Document
-from ginger.libmt94x.fields import AccountIdentification
-from ginger.libmt94x.fields import ClosingAvailableBalance
-from ginger.libmt94x.fields import ClosingBalance
-from ginger.libmt94x.fields import ExportInformation
-from ginger.libmt94x.fields import ForwardAvailableBalance
-from ginger.libmt94x.fields import ImportInformation
-from ginger.libmt94x.fields import InformationToAccountOwner
-from ginger.libmt94x.fields import InformationToAccountOwnerTotals
-from ginger.libmt94x.fields import OpeningBalance
-from ginger.libmt94x.fields import StatementLine
-from ginger.libmt94x.fields import StatementNumber
-from ginger.libmt94x.fields import TransactionReferenceNumber
-from ginger.libmt94x.serializer import Mt94xSerializer
-from ginger.libmt94x.writer import Mt94xWriter
+from ginger_libmt94x.document import Mt940Document
+from ginger_libmt94x.fields import AccountIdentification
+from ginger_libmt94x.fields import ClosingAvailableBalance
+from ginger_libmt94x.fields import ClosingBalance
+from ginger_libmt94x.fields import ExportInformation
+from ginger_libmt94x.fields import ForwardAvailableBalance
+from ginger_libmt94x.fields import ImportInformation
+from ginger_libmt94x.fields import InformationToAccountOwner
+from ginger_libmt94x.fields import InformationToAccountOwnerTotals
+from ginger_libmt94x.fields import OpeningBalance
+from ginger_libmt94x.fields import StatementLine
+from ginger_libmt94x.fields import StatementNumber
+from ginger_libmt94x.fields import TransactionReferenceNumber
+from ginger_libmt94x.serializer import Mt94xSerializer
+from ginger_libmt94x.writer import Mt94xWriter
 
 
 class Mt94xWriterIBPUnstructuredTests(TestCase):