+++ ginger_libmt94x/tests/test_writer_fields.py
@@ -2,28 +2,28 @@ from datetime import datetime
 from decimal import Decimal
 from unittest import TestCase
 
-from ginger.libmt94x.fields import AccountIdentification
-from ginger.libmt94x.fields import ClosingAvailableBalance
-from ginger.libmt94x.fields import ClosingBalance
-from ginger.libmt94x.fields import ExportInformation
-from ginger.libmt94x.fields import ForwardAvailableBalance
-from ginger.libmt94x.fields import InformationToAccountOwner
-from ginger.libmt94x.fields import InformationToAccountOwnerTotals
-from ginger.libmt94x.fields import ImportInformation
-from ginger.libmt94x.fields import OpeningBalance
-from ginger.libmt94x.fields import StatementLine
-from ginger.libmt94x.fields import StatementNumber
-from ginger.libmt94x.fields import TransactionReferenceNumber
-from ginger.libmt94x.info_acct_owner_subfields import CounterPartyID
-from ginger.libmt94x.info_acct_owner_subfields import CreditorID
-from ginger.libmt94x.info_acct_owner_subfields import EndToEndReference
-from ginger.libmt94x.info_acct_owner_subfields import MandateReference
-from ginger.libmt94x.info_acct_owner_subfields import PurposeCode
-from ginger.libmt94x.info_acct_owner_subfields import RemittanceInformation
-from ginger.libmt94x.statement_line_subfields import OriginalAmountOfTransaction
-from ginger.libmt94x.remittance_info import UnstructuredRemittanceInfo
-from ginger.libmt94x.serializer import Mt94xSerializer
-from ginger.libmt94x.writer import Mt94xWriter
+from ginger_libmt94x.fields import AccountIdentification
+from ginger_libmt94x.fields import ClosingAvailableBalance
+from ginger_libmt94x.fields import ClosingBalance
+from ginger_libmt94x.fields import ExportInformation
+from ginger_libmt94x.fields import ForwardAvailableBalance
+from ginger_libmt94x.fields import InformationToAccountOwner
+from ginger_libmt94x.fields import InformationToAccountOwnerTotals
+from ginger_libmt94x.fields import ImportInformation
+from ginger_libmt94x.fields import OpeningBalance
+from ginger_libmt94x.fields import StatementLine
+from ginger_libmt94x.fields import StatementNumber
+from ginger_libmt94x.fields import TransactionReferenceNumber
+from ginger_libmt94x.info_acct_owner_subfields import CounterPartyID
+from ginger_libmt94x.info_acct_owner_subfields import CreditorID
+from ginger_libmt94x.info_acct_owner_subfields import EndToEndReference
+from ginger_libmt94x.info_acct_owner_subfields import MandateReference
+from ginger_libmt94x.info_acct_owner_subfields import PurposeCode
+from ginger_libmt94x.info_acct_owner_subfields import RemittanceInformation
+from ginger_libmt94x.statement_line_subfields import OriginalAmountOfTransaction
+from ginger_libmt94x.remittance_info import UnstructuredRemittanceInfo
+from ginger_libmt94x.serializer import Mt94xSerializer
+from ginger_libmt94x.writer import Mt94xWriter
 
 
 class Mt94xWriterTests(TestCase):